OTA/HttpClient: rework to non-blocking download (select) and generally   applying a read timeout (currently 10 seconds) to avoid blocking   the LwIP API from slow connections and/or high packet losses

+++ vehicle/OVMS.V3/components/ovms_buffer/src/ovms_buffer.h
@@ -60,7 +60,7 @@ class OvmsBuffer
     std::string ReadLine();
 
   public:
-    int PollSocket(int sock, long timeoutms);
+    ssize_t PollSocket(int sock, long timeoutms);
 
   public:
     void* m_userdata;
